Click coordinates made relative.

+++ be/report.go
@@ -12,8 +12,8 @@ type ReportArgs struct {
 	Id       string  `json:"id"`      // public key.
 	Latitude float64 `json:"latitude"`
 	Longitue float64 `json:"longitude"`
-	X        int32   `json:"x"`
-	Y        int32   `json:"y"`
+	X        float64 `json:"x"` // 0.0..1.0
+	Y        float64 `json:"y"` // 0.0..1.0
 	Image    []byte  `json:"image"`
 }
